core/requirements.txt
# Core AI 辨識／轉寫依賴（專案根 core/）
# 從專案根執行 pipeline 時可：pip install -r core/requirements.txt
# 或與 backend 一起裝：pip install -r requirements.txt

python-dotenv
pydub
OpenCC
torch
torchaudio
torchvision
faster-whisper
pyannote-audio
openai
instructor
pydantic

# DIARIZATION_BACKEND=whisper_bilstm 時需 Neuro-AI/ 與下列套件
transformers>=4.36.0
librosa>=0.10.0
soundfile>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
